refactor: 게임 관련 부분 컨트롤러 분리
